Added support for Ooyala videos using hack (sigh!)
Ooyala returns a script that injects the right video player. That
script takes for granted that it is run while the page is loaded
and injects the video player right at the end of the temporary DOM
tree. The problem is that the mediaplayerlib runs when the page is
fully loaded, meaning the video was inserted at the end of the page
no matter what.

There is no good solution to that as long as Ooyala returns that
type of script. The hack being used is to set the global variable
window.ooyalaActiveScript that Ooyala checks and uses to inject the
video player. It works fine but it means that no two Ooyala videos
may be inserted at once for the time being (they would appear in
the same element).

That's probably not a problem for the templates we have so far.
To be checked.