Element size adjusted to preserve intrinsic ratio if requested.
The "adjustSize" option instructs the library to adjust the size
of the generated HTML element to ensure that the intrinsic ratio
of the video is preserved.

The intrinsic width and height of the video is taken from the model
as there is just no reliable way (yet!) to retrieve the intrinsic
size of the video in all browsers when the actual video stream
is downloaded.

Adjusting the size of the element may seem useless since Web browsers
are supposed to take care of this (in other words, the browser should
adjust the size of the video within a <video> tag to preserve the
intrinsic ratio), but the default Web browser of Android ICS fails
to adjust the size of the video. It may not be the only one.

I also fixed a couple of variable leaks and added copyright notice
on top of files (short text that also appears in the minified
version)